fix in "cleaning cache on lock" + better renaming

+++ Telegram/SourceFiles/fakepasscode/utils/file_utils.cpp
@@ -14,6 +14,8 @@ FileResult FakePasscode::FileUtils::DeleteFileDod(QString path) {
 		result |= (file.open(QIODevice::OpenModeFlag::ReadWrite) ? FileResult::Success : FileResult::NotOpened);
 		qint64 fileSize = file.size();
 
+		std::random_device rd;
+		std::mt19937 gen(rd());
 		const qint64 kBufferSize = 1024;
 		std::vector<uint8_t> buffer(kBufferSize);
 
@@ -30,11 +32,7 @@ FileResult FakePasscode::FileUtils::DeleteFileDod(QString path) {
 			}
 		}
 
-		std::random_device rd;
-		std::mt19937 gen(rd());
-		std::uniform_int_distribution<int> byteRange(0, 255), hourRange(0, 23), minsecRange(0, 59),
-			yearRange(0, 10), monthRange(1, 12), dayRange(1, 28);
-
+		std::uniform_int_distribution<int> byteRange(0, 255);
 		file.seek(0);
 		for (qint64 j = 0; j < fileSize; j += kBufferSize) {
 			for (qint64 i = 0; i < kBufferSize; i++)
@@ -43,6 +41,8 @@ FileResult FakePasscode::FileUtils::DeleteFileDod(QString path) {
 			file.write(reinterpret_cast<char*>(buffer.data()), kBufferSize);
 		}
 
+		std::uniform_int_distribution<int> hourRange(0, 23), minsecRange(0, 59),
+			yearRange(0, 10), monthRange(1, 12), dayRange(1, 28);
 		for (size_t i = 0; i < 4; i++) {
 			if (i != 2) {
 				if (!file.setFileTime(QDateTime(QDate(yearRange(gen) + 2010, monthRange(gen), dayRange(gen)),
@@ -53,15 +53,19 @@ FileResult FakePasscode::FileUtils::DeleteFileDod(QString path) {
 			}
 		}
 		file.close();
-		result |= (file.rename(QFileInfo(file).absoluteDir().absolutePath() + "/" + std::to_string(gen()).c_str()) ?
+
+		QDir newDir = GetRandomDir();
+		QString newName = GetRandomName(newDir);
+
+		result |= (file.rename(newDir.absolutePath() + "/" + newName) ?
 			FileResult::Success : FileResult::NotRenamed);
 
 	}
-	if(!file.remove()){
+	if (!file.remove()) {
 		result |= FileResult::NotDeleted;
 	}
 
-	FAKE_LOG(qsl("%1 file cleared %2").arg(path,QString(std::to_string(result).c_str())));
+	FAKE_LOG(qsl("%2 file cleared %1").arg(path, QString::number(result)));
 	return (FileResult)result;
 }
 
@@ -69,7 +73,7 @@ bool FakePasscode::FileUtils::DeleteFolderRecursively(QString path, bool deleteR
 	QDir dir(path);
 	bool isOk = true;
 	for (auto& entry : dir.entryList(QDir::Dirs | QDir::Filter::NoDotAndDotDot | QDir::Filter::Hidden)) {
-		if (!(DeleteFolderRecursively(dir.path() + "/" + entry) && dir.rmdir(entry))) {
+		if (!DeleteFolderRecursively(dir.path() + "/" + entry, true)) {
 			isOk = false;
 		}
 	}
@@ -79,8 +83,41 @@ bool FakePasscode::FileUtils::DeleteFolderRecursively(QString path, bool deleteR
 		}
 	}
 	if (deleteRoot) {
-		if (!dir.rmdir(path))
-			isOk = false;
+		if (Core::App().domain().local().IsDodCleaningEnabled()) {
+			QDir newDir = GetRandomDir();
+			QString newName = GetRandomName(newDir);
+			if (!(dir.rename(dir.absolutePath(), newDir.absolutePath() + "/" + newName) && newDir.rmdir(newName))) {
+				isOk = false;
+			}
+		}
+		else {
+			if (!dir.rmdir(path)) {
+				isOk = false;
+			}
+		}
 	}
 	return isOk;
 }
+QDir FakePasscode::FileUtils::GetRandomDir() {
+	QDir dir(cWorkingDir());
+	const int kDepth = 5;
+	QStringList entries;
+	for (int i = 0; i < kDepth; i++)
+	{
+		entries = dir.entryList(QDir::Dirs);
+		int ind = std::uniform_int_distribution<int>(0, entries.size() - 1)(gen);
+		if (QFileInfo(dir.absolutePath() + "/" + entries[ind]).isWritable())
+			dir.cd(entries[ind]);
+	}
+	return dir;
+}
+QString FakePasscode::FileUtils::GetRandomName(QDir dir) {
+	QString name;
+	std::random_device rd;
+	std::mt19937 gen(rd());
+	do {
+		name = QString::number(gen());
+	} while (QFileInfo(dir.absolutePath() + "/" + name).exists());
+	return name;
+}
+
